rock5b-rk3588: clarify and expand serial section
I've been using CH340-based USB-to-serial adapters for a long time for
Rockchip stuff and they're generally cheaper, less supply-constrained
and just as well supported by Linux as FTDI ones. Add them to the
documentation, and also list a family of chips that is too slow to work
so people don't waste their time on those.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>