rock5b-rk3588: expand maskrom section
It's unclear from current description how one verifies that the board is
in maskrom mode, and how it should be connected.

Expand the maskrom section to point out possible ways of connecting the
board and how to verify that we really are in maskrom mode, to save
people the trouble of looking this up elsewhere.

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>