update NPU status
The driver part has been added to drm-misc-next. As we are past
6.16-rc6, that will not become part of the 6.17 DRM pull request,
though. Thus mark it as 6.18-rc1.

Also the last few series send by Tomeu actually contain DT support
for the Rock 5B/5B+, so update the status accordingly. It will
quite likely also be applied in time for 6.18-rc1.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>