Add tracking link for HDMI PHY FRL support
HDMI Fixed Rate Link (FRL) is needed for resolutions above 4K 60 FPS. It
is part of the HDMI 2.1 spec and works quite differently from the TMDS
mode used by previous standards.

NOTE: In addition to the HDMI PHY, the controller also needs to be
updated. This will be handled separately.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>