refine and update camera and video capture
The camera and video capture part consists of many (sub)components:
 - MIPI CSI PHYs (two types of them: DPHY and combo DCPHY)
 - ISP
 - VICAP, including DVP, MIPI CSI Hosts, MUX/TOISP, SCALER
Add them to the status matrix in order to present a granular overview.

The blocks are typically enabled only if external hardware is attached
to a board, e.g., a camera module. Thus, there is no work to be done
on the board DTS level. Reflect that with n/a in the board colums.

While at it, update the links to the latest versions.