RK3568 VICAP got merged
While RK3568 VICAP is not RK3588 VICAP, there is a lot of shared code.
This means that the RK3568 was more or less a perquisite for supporting
RK3588's VICAP. The patches have not yet been send out, but Michael is
already working on extending the driver for usage on RK3588. A preview
of those patches can be found in our rockchip-devel branch.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>