add link for broken IRQ detection in HDMI-RX
The vendor TF-A for RK3588 grabs the HDMI-RX interrupt, so it cannot be
used directly. Instead it acts as a proxy and provides the functionality
on one of the reserved interrupt lines. The vendor RK3588 device-tree's
use this 'fake' interrupt for HDMI-RX.

The upstream TF-A does not have this proxy functionality (which supposedly
works around some bug in the HDMI-RX IP) and the upstream kernel DT use
the real interrupt from the HDMI-RX.

The result is, that the upstream kernel requires using the upstream TF-A
for HDMI-RX functionality. Using the vendor TF-A results in non-working
HDMI-RX. The patch from Dmitry warns, if such a situation is detected.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>